Updated Auto Paths/ new PathGroup Config
Changed some of the path to move so April tags were in view quicker. This needs to be checked by Aidan to make sure the paths look good. I also added the new calling of PathGroups. It works as intended in Robot Simulation but will need to be tested with actual robot to see for sure.